+++ src/main/java/com/kdt/dto/EstateDTO.java
@@ -8,12 +8,12 @@ public class EstateDTO {
     private String writer;
     private Long deposit;
     private Long price;
-    private float area;
+    private double area;
     private Long zipcode;
     private String address1;
     private String address2;
-    private float latitude;
-    private float longitude;
+    private double latitude;
+    private double longitude;
     private Long roomFloors;
     private Long buildingFloors;
     private Long maintenanceCost;
@@ -28,7 +28,6 @@ public class EstateDTO {
     private HeatingSystemDTO heatingSystem;
     private Set<EstateOptionDTO> optionList;
     private Set<EstateImageDTO> images;
-    
 	public Long getEstateId() {
 		return estateId;
 	}
@@ -53,10 +52,10 @@ public class EstateDTO {
 	public void setPrice(Long price) {
 		this.price = price;
 	}
-	public float getArea() {
+	public double getArea() {
 		return area;
 	}
-	public void setArea(float area) {
+	public void setArea(double area) {
 		this.area = area;
 	}
 	public Long getZipcode() {
@@ -77,16 +76,16 @@ public class EstateDTO {
 	public void setAddress2(String address2) {
 		this.address2 = address2;
 	}
-	public float getLatitude() {
+	public double getLatitude() {
 		return latitude;
 	}
-	public void setLatitude(float latitude) {
+	public void setLatitude(double latitude) {
 		this.latitude = latitude;
 	}
-	public float getLongitude() {
+	public double getLongitude() {
 		return longitude;
 	}
-	public void setLongitude(float longitude) {
+	public void setLongitude(double longitude) {
 		this.longitude = longitude;
 	}
 	public Long getRoomFloors() {
@@ -173,8 +172,8 @@ public class EstateDTO {
 	public void setImages(Set<EstateImageDTO> images) {
 		this.images = images;
 	}
-	public EstateDTO(Long estateId, String writer, Long deposit, Long price, float area, Long zipcode, String address1,
-			String address2, float latitude, float longitude, Long roomFloors, Long buildingFloors,
+	public EstateDTO(Long estateId, String writer, Long deposit, Long price, double area, Long zipcode, String address1,
+			String address2, double latitude, double longitude, Long roomFloors, Long buildingFloors,
 			Long maintenanceCost, String title, String contents, String memo, Timestamp writeDate, RoomDTO room,
 			StructureDTO structure, BuildingDTO building, TransactionDTO transaction, HeatingSystemDTO heatingSystem,
 			Set<EstateOptionDTO> optionList, Set<EstateImageDTO> images) {
@@ -207,4 +206,6 @@ public class EstateDTO {
 	public EstateDTO() {
 		super();
 	}
+    
+	
 }
